+++ src/lambda/expression.rs
@@ -1,17 +1,15 @@
-use std::collections::HashSet;
 use std::fmt::Debug;
 use std::sync::Arc;
 
 use anyhow::Result;
 use async_graphql_value::ConstValue;
-use futures_util::future::join_all;
-use futures_util::stream::FuturesUnordered;
-use futures_util::StreamExt;
 use reqwest::Request;
 use serde_json::Value;
 use thiserror::Error;
 
-use super::{eval_math, Eval, Math, Relation, ResolverContextLike};
+use super::list::List;
+use super::logic::Logic;
+use super::{Eval, Math, Relation, ResolverContextLike};
 use crate::config::group_by::GroupBy;
 use crate::config::GraphQLOperationType;
 use crate::data_loader::{DataLoader, Loader};
@@ -21,7 +19,6 @@ use crate::grpc::data_loader::GrpcDataLoader;
 use crate::grpc::protobuf::ProtobufOperation;
 use crate::grpc::request::execute_grpc_request;
 use crate::grpc::request_template::RenderedRequestTemplate;
-use crate::helpers::value::HashableConstValue;
 use crate::http::{self, cache_policy, DataLoaderRequest, HttpDataLoader, Response};
 #[cfg(feature = "unsafe-js")]
 use crate::javascript;
@@ -46,26 +43,6 @@ pub enum Expression {
 pub enum Concurrency {
   Parallel,
   Sequential,
-}
-
-#[derive(Clone, Debug)]
-pub enum List {
-  Concat(Vec<Expression>),
-}
-
-#[derive(Clone, Debug)]
-pub enum Logic {
-  If {
-    cond: Box<Expression>,
-    then: Box<Expression>,
-    els: Box<Expression>,
-  },
-  And(Vec<Expression>),
-  Or(Vec<Expression>),
-  Cond(Box<Expression>, Vec<(Box<Expression>, Box<Expression>)>),
-  DefaultTo(Box<Expression>, Box<Expression>),
-  IsEmpty(Box<Expression>),
-  Not(Box<Expression>),
 }
 
 #[derive(Clone, Debug)]
@@ -250,212 +227,9 @@ impl Eval for Expression {
       Expression::Relation(relation) => relation.async_eval(ctx, conc).await,
       Expression::Logic(logic) => logic.async_eval(ctx, conc).await,
       Expression::List(list) => list.async_eval(ctx, conc).await,
-      Expression::Math(math) => eval_math(ctx, math, conc).await,
+      Expression::Math(math) => math.async_eval(ctx, conc).await,
     }
   }
-}
-
-impl Eval for List {
-  async fn async_eval<'a, Ctx: ResolverContextLike<'a> + Sync + Send>(
-    &'a self,
-    ctx: &'a EvaluationContext<'a, Ctx>,
-    conc: &'a Concurrency,
-  ) -> Result<async_graphql::Value> {
-    match self {
-      List::Concat(list) => join_all(list.iter().map(|expr| expr.eval(ctx, conc)))
-        .await
-        .into_iter()
-        .try_fold(async_graphql::Value::List(vec![]), |acc, result| match (acc, result?) {
-          (ConstValue::List(mut lhs), ConstValue::List(rhs)) => {
-            lhs.extend(rhs.into_iter());
-            Ok(ConstValue::List(lhs))
-          }
-          _ => Err(EvaluationError::ConcatException("element is not a list".into()))?,
-        }),
-    }
-  }
-}
-
-impl Eval for Logic {
-  async fn async_eval<'a, Ctx: ResolverContextLike<'a> + Sync + Send>(
-    &'a self,
-    ctx: &'a EvaluationContext<'a, Ctx>,
-    conc: &'a Concurrency,
-  ) -> Result<async_graphql::Value> {
-    Ok(match self {
-      Logic::Or(list) => {
-        let future_iter = list.iter().map(|expr| expr.eval(ctx, conc));
-
-        match *conc {
-          Concurrency::Parallel => {
-            let mut futures: FuturesUnordered<_> = future_iter.collect();
-            let mut output = false;
-
-            while let Some(result) = futures.next().await {
-              let result: Result<ConstValue> = result;
-              if is_truthy(result?) {
-                output = true;
-                break;
-              }
-            }
-            output
-          }
-          Concurrency::Sequential => {
-            let mut output = false;
-            for future in future_iter.into_iter() {
-              if is_truthy(future.await?) {
-                output = true;
-                break;
-              }
-            }
-            output
-          }
-        }
-        .into()
-      }
-      Logic::Cond(default, list) => match *conc {
-        Concurrency::Sequential => {
-          let mut result = None;
-          for (cond, expr) in list.iter() {
-            if is_truthy(cond.eval(ctx, conc).await?) {
-              result = Some(expr.eval(ctx, conc).await?);
-              break;
-            }
-          }
-          result.unwrap_or(default.eval(ctx, conc).await?)
-        }
-        Concurrency::Parallel => {
-          let true_cond_index = join_all(list.iter().map(|(cond, _)| cond.eval(ctx, conc)))
-            .await
-            .into_iter()
-            .enumerate()
-            .find_map(|(index, cond)| Some(is_truthy_ref(cond.as_ref().ok()?)).map(|_| index));
-
-          if let Some(index) = true_cond_index {
-            let (_, value) = list
-              .get(index)
-              .ok_or(anyhow::anyhow!("no condition found at index {index}"))?;
-            value.eval(ctx, conc).await?
-          } else {
-            default.eval(ctx, conc).await?
-          }
-        }
-      },
-      Logic::DefaultTo(value, default) => {
-        let result = value.eval(ctx, conc).await?;
-        if is_empty(&result) {
-          default.eval(ctx, conc).await?
-        } else {
-          result
-        }
-      }
-      Logic::IsEmpty(expr) => is_empty(&expr.eval(ctx, conc).await?).into(),
-      Logic::Not(expr) => (!is_truthy(expr.eval(ctx, conc).await?)).into(),
-
-      Logic::And(list) => {
-        let future_iter = list.iter().map(|expr| expr.eval(ctx, conc));
-
-        match *conc {
-          Concurrency::Parallel => {
-            let mut futures: FuturesUnordered<_> = future_iter.collect();
-            let mut output = true;
-
-            while let Some(result) = futures.next().await {
-              let result: Result<ConstValue> = result;
-              if !is_truthy(result?) {
-                output = false;
-                break;
-              }
-            }
-            output
-          }
-          Concurrency::Sequential => {
-            let mut output = true;
-            for future in future_iter.into_iter() {
-              if !is_truthy(future.await?) {
-                output = false;
-                break;
-              }
-            }
-            output
-          }
-        }
-        .into()
-      }
-      Logic::If { cond, then, els } => {
-        let cond = cond.eval(ctx, conc).await?;
-        if is_truthy(cond) {
-          then.eval(ctx, conc).await?
-        } else {
-          els.eval(ctx, conc).await?
-        }
-      }
-    })
-  }
-}
-
-pub async fn eval_list_expressions<'a, Ctx: ResolverContextLike<'a> + Sync + Send, C: FromIterator<ConstValue>>(
-  ctx: &'a EvaluationContext<'a, Ctx>,
-  conc: &'a Concurrency,
-  exprs: &'a [Expression],
-) -> Result<C> {
-  let future_iter = exprs.iter().map(|expr| expr.eval(ctx, conc));
-  match *conc {
-    Concurrency::Parallel => join_all(future_iter).await.into_iter().collect::<Result<C>>(),
-    Concurrency::Sequential => {
-      let mut results = Vec::with_capacity(exprs.len());
-      for future in future_iter {
-        results.push(future.await?);
-      }
-      Ok(results.into_iter().collect())
-    }
-  }
-}
-
-#[allow(clippy::redundant_closure, clippy::too_many_arguments)]
-async fn eval_map_list_expressions<
-  'a,
-  Ctx: ResolverContextLike<'a> + Sync + Send,
-  O,
-  C: FromIterator<O>,
-  F: Fn(ConstValue) -> O,
->(
-  ctx: &'a EvaluationContext<'a, Ctx>,
-  conc: &'a Concurrency,
-  exprs: &'a [Expression],
-  f: F,
-) -> Result<C> {
-  let future_iter = exprs.iter().map(|expr| expr.eval(ctx, conc));
-  match *conc {
-    Concurrency::Parallel => join_all(future_iter)
-      .await
-      .into_iter()
-      .map(|result| result.map(|cv| f(cv)))
-      .collect::<Result<C>>(),
-    Concurrency::Sequential => {
-      let mut results = Vec::with_capacity(exprs.len());
-      for future in future_iter {
-        results.push(f(future.await?));
-      }
-      Ok(results.into_iter().collect())
-    }
-  }
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn set_operation<'a, 'b, Ctx: ResolverContextLike<'a> + Sync + Send, F>(
-  ctx: &'a EvaluationContext<'a, Ctx>,
-  conc: &'a Concurrency,
-  lhs: &'a [Expression],
-  rhs: &'a [Expression],
-  operation: F,
-) -> Result<ConstValue>
-where
-  F: Fn(HashSet<HashableConstValue>, HashSet<HashableConstValue>) -> Vec<ConstValue>,
-{
-  let lhs = eval_map_list_expressions(ctx, conc, lhs, HashableConstValue).await?;
-  let rhs = eval_map_list_expressions(ctx, conc, rhs, HashableConstValue).await?;
-  Ok(operation(lhs, rhs).into())
 }
 
 pub fn get_path_for_const_value_owned(path: &[impl AsRef<str>], mut const_value: ConstValue) -> Option<ConstValue> {
@@ -481,54 +255,6 @@ pub fn get_path_for_const_value_ref<'a>(
   }
 
   Some(const_value)
-}
-
-/// Check if a value is truthy
-///
-/// Special cases:
-/// 1. An empty string is considered falsy
-/// 2. A collection of bytes is truthy, even if the value in those bytes is 0. An empty collection is falsy.
-fn is_truthy(value: async_graphql::Value) -> bool {
-  use async_graphql::{Number, Value};
-  use hyper::body::Bytes;
-
-  match value {
-    Value::Null => false,
-    Value::Enum(_) => true,
-    Value::List(_) => true,
-    Value::Object(_) => true,
-    Value::String(s) => !s.is_empty(),
-    Value::Boolean(b) => b,
-    Value::Number(n) => n != Number::from(0),
-    Value::Binary(b) => b != Bytes::default(),
-  }
-}
-
-fn is_truthy_ref(value: &async_graphql::Value) -> bool {
-  use async_graphql::{Number, Value};
-  use hyper::body::Bytes;
-
-  match value {
-    &Value::Null => false,
-    &Value::Enum(_) => true,
-    &Value::List(_) => true,
-    &Value::Object(_) => true,
-    Value::String(s) => !s.is_empty(),
-    &Value::Boolean(b) => b,
-    Value::Number(n) => n != &Number::from(0),
-    Value::Binary(b) => b != &Bytes::default(),
-  }
-}
-
-fn is_empty(value: &async_graphql::Value) -> bool {
-  match value {
-    ConstValue::Null => true,
-    ConstValue::Number(_) | ConstValue::Boolean(_) | ConstValue::Enum(_) => false,
-    ConstValue::Binary(bytes) => bytes.is_empty(),
-    ConstValue::List(list) => list.is_empty(),
-    ConstValue::Object(obj) => obj.is_empty(),
-    ConstValue::String(string) => string.is_empty(),
-  }
 }
 
 fn set_cache_control<'ctx, Ctx: ResolverContextLike<'ctx>>(
@@ -643,7 +369,7 @@ mod tests {
   use hyper::body::Bytes;
   use indexmap::IndexMap;
 
-  use super::is_truthy;
+  use crate::lambda::is_truthy;
 
   #[test]
   fn test_is_truthy() {
