+++ src/lambda/mod.rs
@@ -2,7 +2,10 @@ mod eval;
 mod evaluation_context;
 mod expression;
 mod graphql_operation_context;
+mod helpers;
 mod lambda;
+mod list;
+mod logic;
 mod math;
 mod relation;
 mod resolver_context_like;
@@ -11,7 +14,10 @@ pub use eval::*;
 pub use evaluation_context::EvaluationContext;
 pub(crate) use expression::*;
 pub use graphql_operation_context::GraphQLOperationContext;
+pub use helpers::*;
 pub use lambda::Lambda;
+pub use list::*;
+pub use logic::*;
 pub use math::*;
 pub use relation::*;
 pub use resolver_context_like::{EmptyResolverContext, ResolverContextLike};
